type fix for max_emb_index

+++ src/faceID.c
@@ -537,7 +537,7 @@ static void run_cnn_2(int x_offset, int y_offset)
 
     int32_t *ml_point =  ml_3_data32;
     int8_t max_emb = 0;
-    int8_t max_emb_index = 0;
+    int32_t max_emb_index = 0;
     char* name;
        
      for (int i = 0; i < (DEFAULT_EMBS_NUM + db_flash_emb_count + 3) / 4; i++)
